In sx126x, swap RxTimeout check with Preamble detection check.
With poor timing, or a busy CPU, the radio can signal Timeout & PreambleDetect in the same IrqStatus read.
The current code will return PreambleDetect in this case, and then happily waits for the timeout event.
This results in a infinite stall where the upper level is waiting for a timeout event that has already arrived, manifesting in receive(RxSingle) never timing out.
By swaping the preamble detection with the timeout, now the timeout will trigger in that case, which will correctly be propagated up the stack to the calling functions as a timeout.